sketch 0.0.5 { information-service. information/place/{restaurantID} }

+++ information-service/src/main/java/com/beta/informationservice/model/Place.java
@@ -10,10 +10,10 @@ public class Place {
 
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
-    private int placeID;
+    private long placeID;
     private String name;
 
-    public int getPlaceID() {
+    public long getPlaceID() {
         return placeID;
     }
